igw: don't overwrite iscsi-gateway.cfg on rerurns
It is common for users to deply a setup, modify the iscsi group vars,
then rerun the ansible playbook command. This will end up overwriting
the iscsi-gateway.cfg.

This has use use the force no option so the cfg file is only copied if
the file does not yet exist.