ceph-iscsi-gw: drop rtslib from prereqs
rtslib with the needed changes now ships in RHEL/Centos 7.7. This drops
it from the prerequisites.yml. We then rely on the ceph-iscsi and targetcli
rpm specs to list the rtslib depedency and bring in the correct version.

This fixes the bug where when using python3 like default on RHEL 8 the
name of the package is python3-rtslib, so installation will fail trying
to bring in python-rtslib.

Signed-off-by: Mike Christie <mchristi@redhat.com>